Renamed fetchObjects to fetchEvents
