+++ nodes/CalDavNode/actions/calendarObject/fetchEvents/description.ts
@@ -14,7 +14,7 @@ export const description: CalendarObjectProperties = [
 		displayOptions: {
 			show: {
 				resource: ['calendarObject'],
-				operation: ['fetchObjects']
+				operation: ['fetchEvents']
 			}
 		},
 		description: 'Calendar to work with. Choose from the list, or specify an URL using an <a href="https://docs.n8n.io/code-examples/expressions/">expression</a>. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code-examples/expressions/">expression</a>.',
